Text Embeddings & RAGs /requirements.txt
pdfplumber>=0.11.0
chromadb>=0.5.0
sentence-transformers>=3.0.0
anthropic[bedrock]>=0.40.0
python-dotenv>=1.0.0
